Rotations are now doubles. World transform properties can now be set! I think I did the math right?

+++ SkiEngine/Extensions/SKCanvasExtensions.cs
@@ -20,7 +20,7 @@ namespace SkiEngine.Extensions
             if (rotated)
             {
                 canvas.Save();
-                canvas.RotateRadians(transform.WorldRotation);
+                canvas.RotateRadians((float)transform.WorldRotation);
             }
             
             canvas.DrawImage(
