+++ SkiEngine/Interfaces/ITransform.cs
@@ -5,7 +5,7 @@ namespace SkiEngine.Interfaces
     public interface ITransform
     {
         SKPoint WorldPoint { get; set; }
-        float WorldRotation { get; set; }
+        double WorldRotation { get; set; }
         SKPoint WorldScale { get; set; }
     }
 }
